requirements.txt
streamlit
pillow
pytesseract
rapidfuzz
groq
python-dotenv
